Fix adding world levels if region has no objects dir (LoA)

+++ world_levels.py
@@ -47,17 +47,20 @@ def copy_wl_files(region: Region):
     os.remove(os.path.join(index_dir.path, 'streamer_node_content_index.gas'))
 
     objects_dir = region.gas_dir.get_subdir('objects')
-    for wl in ['regular', 'veteran', 'elite']:
-        os.mkdir(os.path.join(objects_dir.path, wl))
+    if objects_dir is None:
+        print(f'  {region.get_name()} has no objects dir')
+    else:
+        for wl in ['regular', 'veteran', 'elite']:
+            os.mkdir(os.path.join(objects_dir.path, wl))
+            for file_name in os.listdir(objects_dir.path):
+                if not file_name.endswith('.gas'):
+                    continue
+                shutil.copy(os.path.join(objects_dir.path, file_name), os.path.join(objects_dir.path, wl, file_name))
+            time.sleep(0.1)  # shutil...
         for file_name in os.listdir(objects_dir.path):
             if not file_name.endswith('.gas'):
                 continue
-            shutil.copy(os.path.join(objects_dir.path, file_name), os.path.join(objects_dir.path, wl, file_name))
-        time.sleep(0.1)  # shutil...
-    for file_name in os.listdir(objects_dir.path):
-        if not file_name.endswith('.gas'):
-            continue
-        os.remove(os.path.join(objects_dir.path, file_name))
+            os.remove(os.path.join(objects_dir.path, file_name))
 
 
 def adapt_file_templates(wl_dir: GasDir, wl_prefix: str, file_name: str, static_template_names: list[str], existing_template_names: list[str]):
@@ -116,6 +119,9 @@ def adapt_condition_params(wl_dir: GasDir, wl_prefix: str, actor_template_names:
 
 def adapt_templates(region: Region, static_template_names: dict[str, list[str]], existing_template_names: dict[str, list[str]]):
     objects_dir = region.gas_dir.get_subdir('objects')
+    if objects_dir is None:
+        print(f'  {region.get_name()} has no objects dir')
+        return
     for wl, prefix in {'veteran': '2W_', 'elite': '3W_'}.items():
         wl_dir = objects_dir.get_subdir(wl)
         adapt_file_templates(wl_dir, prefix, 'actor', static_template_names['core'], existing_template_names['actor'])
@@ -138,6 +144,9 @@ def is_to_delete_for_wl(go_section: Section, wl: str):
 def delete_tutorial_tips(region: Region):
     index_dir = region.gas_dir.get_subdir('index')
     objects_dir = region.gas_dir.get_subdir('objects')
+    if objects_dir is None:
+        print(f'  {region.get_name()} has no objects dir')
+        return
     for wl, prefix in {'veteran': '2W_', 'elite': '3W_'}.items():
         wl_dir = objects_dir.get_subdir(wl)
         if wl_dir.has_gas_file('special'):
@@ -204,6 +213,9 @@ def scale_shrine(shrine_section: Section, wl: str):
 
 def scale_shrines(region: Region):
     objects_dir = region.gas_dir.get_subdir('objects')
+    if objects_dir is None:
+        print(f'  {region.get_name()} has no objects dir')
+        return
     for wl, prefix in {'veteran': '2W_', 'elite': '3W_'}.items():
         wl_dir = objects_dir.get_subdir(wl)
         if wl_dir.has_gas_file('special'):
